+++ src/mjolnir/util.cc
@@ -22,6 +22,21 @@
 #include <boost/filesystem/operations.hpp>
 #include <boost/property_tree/ptree.hpp>
 
+namespace {
+
+// Temporary files used during tile building
+const std::string ways_file = "ways.bin";
+const std::string way_nodes_file = "way_nodes.bin";
+const std::string nodes_file = "nodes.bin";
+const std::string edges_file = "edges.bin";
+const std::string access_file = "access.bin";
+const std::string cr_from_file = "complex_from_restrictions.bin";
+const std::string cr_to_file = "complex_to_restrictions.bin";
+const std::string new_to_old_file = "new_nodes_to_old_nodes.bin";
+const std::string old_to_new_file = "old_nodes_to_new_nodes.bin";
+
+} // namespace
+
 namespace valhalla {
 namespace mjolnir {
 
@@ -112,75 +127,104 @@ bool shapes_match(const std::vector<PointLL>& shape1, const std::vector<PointLL>
 
 void build_tile_set(const boost::property_tree::ptree& config,
                     const std::vector<std::string>& input_files,
-                    const std::string& bin_file_prefix,
-                    bool free_protobuf) {
+                    const BuildStage start_stage,
+                    const BuildStage end_stage) {
+  auto remove_temp_file = [](const std::string& fname) {
+    if (boost::filesystem::exists(fname)) {
+      boost::filesystem::remove(fname);
+    }
+  };
+
   // cannot allow this when building tiles
   if (config.get_child("mjolnir").get_optional<std::string>("tile_extract")) {
     throw std::runtime_error("Tiles cannot be directly built into a tar extract");
   }
 
-  // set up the directories and purge old tiles
-  auto tile_dir = config.get<std::string>("mjolnir.tile_dir");
-  for (const auto& level : valhalla::baldr::TileHierarchy::levels()) {
-    auto level_dir = tile_dir + filesystem::path::preferred_separator + std::to_string(level.first);
+  // Get the tile directory (make sure it ends with the preferred separator
+  std::string tile_dir = config.get<std::string>("mjolnir.tile_dir");
+  if (tile_dir.back() != filesystem::path::preferred_separator) {
+    tile_dir.push_back(filesystem::path::preferred_separator);
+  }
+
+  // Set up the temporary (*.bin) files used during processing
+  std::string ways_bin = tile_dir + ways_file;
+  std::string way_nodes_bin = tile_dir + way_nodes_file;
+  std::string nodes_bin = tile_dir + nodes_file;
+  std::string edges_bin = tile_dir + edges_file;
+  std::string access_bin = tile_dir + access_file;
+  std::string cr_from_bin = tile_dir + cr_from_file;
+  std::string cr_to_bin = tile_dir + cr_to_file;
+  std::string new_to_old_bin = tile_dir + new_to_old_file;
+  std::string old_to_new_bin = tile_dir + old_to_new_file;
+
+  // Read the OSM protocol buffer file. Callbacks for nodes, ways, and
+  // relations are defined within the PBFParser class
+  // NOTE: parse and build stages must be run together due to osm_data
+  if (start_stage == BuildStage::kParse && BuildStage::kBuild <= end_stage) {
+    // set up the directories and purge old tiles if starting at the parsing stage
+    for (const auto& level : valhalla::baldr::TileHierarchy::levels()) {
+      auto level_dir = tile_dir + std::to_string(level.first);
+      if (boost::filesystem::exists(level_dir) && !boost::filesystem::is_empty(level_dir)) {
+        LOG_WARN("Non-empty " + level_dir + " will be purged of tiles");
+        boost::filesystem::remove_all(level_dir);
+      }
+    }
+
+    // check for transit level.
+    auto level_dir =
+        tile_dir +
+        std::to_string(valhalla::baldr::TileHierarchy::levels().rbegin()->second.level + 1);
     if (boost::filesystem::exists(level_dir) && !boost::filesystem::is_empty(level_dir)) {
       LOG_WARN("Non-empty " + level_dir + " will be purged of tiles");
       boost::filesystem::remove_all(level_dir);
     }
-  }
 
-  // check for transit level.
-  auto level_dir =
-      tile_dir + filesystem::path::preferred_separator +
-      std::to_string(valhalla::baldr::TileHierarchy::levels().rbegin()->second.level + 1);
-  if (boost::filesystem::exists(level_dir) && !boost::filesystem::is_empty(level_dir)) {
-    LOG_WARN("Non-empty " + level_dir + " will be purged of tiles");
-    boost::filesystem::remove_all(level_dir);
-  }
+    boost::filesystem::create_directories(tile_dir);
 
-  boost::filesystem::create_directories(tile_dir);
+    // Parse OSM data
+    auto osm_data = PBFGraphParser::Parse(config.get_child("mjolnir"), input_files, ways_bin,
+                                          way_nodes_bin, access_bin, cr_from_bin, cr_to_bin);
 
-  // Read the OSM protocol buffer file. Callbacks for nodes, ways, and
-  // relations are defined within the PBFParser class
-  auto osm_data =
-      PBFGraphParser::Parse(config.get_child("mjolnir"), input_files, bin_file_prefix + "ways.bin",
-                            bin_file_prefix + "way_nodes.bin", bin_file_prefix + "access.bin",
-                            bin_file_prefix + "complex_from_restrictions.bin",
-                            bin_file_prefix + "complex_to_restrictions.bin");
-
-  // Optionally free all protobuf memory but also you cant use the protobuffer lib after this!
-  if (free_protobuf) {
+    // Free all protobuf memory - cannot use the protobuffer lib after this!
     OSMPBF::Parser::free();
-  }
 
-  // Build the graph using the OSMNodes and OSMWays from the parser
-  GraphBuilder::Build(config, osm_data, bin_file_prefix + "ways.bin",
-                      bin_file_prefix + "way_nodes.bin",
-                      bin_file_prefix + "complex_from_restrictions.bin",
-                      bin_file_prefix + "complex_to_restrictions.bin");
+    // Build the graph using the OSMNodes and OSMWays from the parser
+    GraphBuilder::Build(config, osm_data, ways_bin, way_nodes_bin, nodes_bin, edges_bin, cr_from_bin,
+                        cr_to_bin);
+  }
 
   // Enhance the local level of the graph. This adds information to the local
   // level that is usable across all levels (density, administrative
   // information (and country based attribution), edge transition logic, etc.
-  GraphEnhancer::Enhance(config, bin_file_prefix + "access.bin");
+  if (start_stage <= BuildStage::kEnhance && BuildStage::kEnhance <= end_stage) {
+    GraphEnhancer::Enhance(config, access_bin);
+  }
 
   // Perform optional edge filtering (remove edges and nodes for specific access modes)
-  GraphFilter::Filter(config);
+  if (start_stage <= BuildStage::kFilter && BuildStage::kFilter <= end_stage) {
+    GraphFilter::Filter(config);
+  }
 
   // Add transit
-  TransitBuilder::Build(config);
+  if (start_stage <= BuildStage::kTransit && BuildStage::kTransit <= end_stage) {
+    TransitBuilder::Build(config);
+  }
 
   // Builds additional hierarchies if specified within config file. Connections
   // (directed edges) are formed between nodes at adjacent levels.
   auto build_hierarchy = config.get<bool>("mjolnir.hierarchy", true);
   if (build_hierarchy) {
-    HierarchyBuilder::Build(config);
+    if (start_stage <= BuildStage::kHierarchy && BuildStage::kHierarchy <= end_stage) {
+      HierarchyBuilder::Build(config, new_to_old_bin, old_to_new_bin);
+    }
 
     // Build shortcuts if specified in the config file. Shortcuts can only be
     // applied if hierarchies are also generated.
     auto build_shortcuts = config.get<bool>("mjolnir.shortcuts", true);
     if (build_shortcuts) {
-      ShortcutBuilder::Build(config);
+      if (start_stage <= BuildStage::kShortcuts && BuildStage::kShortcuts <= end_stage) {
+        ShortcutBuilder::Build(config);
+      }
     } else {
       LOG_INFO("Skipping shortcut builder");
     }
@@ -189,12 +233,28 @@ void build_tile_set(const boost::property_tree::ptree& config,
   }
 
   // Build the Complex Restrictions
-  RestrictionBuilder::Build(config, bin_file_prefix + "complex_from_restrictions.bin",
-                            "complex_to_restrictions.bin");
+  if (start_stage <= BuildStage::kRestrictions && BuildStage::kRestrictions <= end_stage) {
+    RestrictionBuilder::Build(config, cr_from_bin, cr_to_bin);
+  }
 
-  // Validate the graph and add information that cannot be added until
-  // full graph is formed.
-  GraphValidator::Validate(config);
+  // Validate the graph and add information that cannot be added until full graph is formed.
+  if (start_stage <= BuildStage::kValidate && BuildStage::kValidate <= end_stage) {
+    GraphValidator::Validate(config);
+  }
+
+  // Cleanup bin files
+  if (start_stage <= BuildStage::kCleanup && BuildStage::kCleanup <= end_stage) {
+    LOG_INFO("Cleaning up temporary *.bin files within " + tile_dir);
+    remove_temp_file(ways_bin);
+    remove_temp_file(way_nodes_bin);
+    remove_temp_file(nodes_bin);
+    remove_temp_file(edges_bin);
+    remove_temp_file(access_bin);
+    remove_temp_file(cr_from_bin);
+    remove_temp_file(cr_to_bin);
+    remove_temp_file(new_to_old_bin);
+    remove_temp_file(old_to_new_bin);
+  }
 }
 
 } // namespace mjolnir