+++ valhalla/mjolnir/util.h
@@ -4,11 +4,60 @@
 #include <boost/property_tree/ptree.hpp>
 #include <list>
 #include <string>
-#include <valhalla/midgard/pointll.h>
+#include <unordered_map>
 #include <vector>
+
+#include <valhalla/midgard/pointll.h>
 
 namespace valhalla {
 namespace mjolnir {
+
+// NOTE - for now parse and build must be run at the same time!
+// Stages of the Valhalla tile building pipeline
+enum class BuildStage : int8_t {
+  kInvalid = -1,
+  kParse = 0,
+  kBuild = 1,
+  kEnhance = 2,
+  kFilter = 3,
+  kTransit = 4,
+  kHierarchy = 5,
+  kShortcuts = 6,
+  kRestrictions = 7,
+  kValidate = 8,
+  kCleanup = 9
+};
+
+// Convert string to BuildStage
+inline BuildStage string_to_buildstage(const std::string& s) {
+  static const std::unordered_map<std::string, BuildStage> stringToBuildStage =
+      {{"parse", BuildStage::kParse},         {"build", BuildStage::kBuild},
+       {"enhance", BuildStage::kEnhance},     {"filter", BuildStage::kFilter},
+       {"transit", BuildStage::kTransit},     {"hierarchy", BuildStage::kHierarchy},
+       {"shortcuts", BuildStage::kShortcuts}, {"restrictions", BuildStage::kRestrictions},
+       {"validate", BuildStage::kValidate},   {"cleanup", BuildStage::kCleanup}};
+
+  auto i = stringToBuildStage.find(s);
+  return (i == stringToBuildStage.cend()) ? BuildStage::kInvalid : i->second;
+}
+
+// Convert BuildStage to string
+inline std::string to_string(BuildStage stg) {
+  static const std::unordered_map<uint8_t, std::string> BuildStageStrings =
+      {{static_cast<int8_t>(BuildStage::kParse), "parse"},
+       {static_cast<int8_t>(BuildStage::kBuild), "build"},
+       {static_cast<int8_t>(BuildStage::kEnhance), "enhance"},
+       {static_cast<int8_t>(BuildStage::kFilter), "filter"},
+       {static_cast<int8_t>(BuildStage::kTransit), "transit"},
+       {static_cast<int8_t>(BuildStage::kHierarchy), "hierarchy"},
+       {static_cast<int8_t>(BuildStage::kShortcuts), "shortcuts"},
+       {static_cast<int8_t>(BuildStage::kRestrictions), "restrictions"},
+       {static_cast<int8_t>(BuildStage::kValidate), "validate"},
+       {static_cast<int8_t>(BuildStage::kCleanup), "cleanup"}};
+
+  auto i = BuildStageStrings.find(static_cast<int8_t>(stg));
+  return (i == BuildStageStrings.cend()) ? "null" : i->second;
+}
 
 /**
  * Splits a tag into a vector of strings.
@@ -43,18 +92,18 @@ bool shapes_match(const std::vector<midgard::PointLL>& shape1,
 uint32_t compute_curvature(const std::list<midgard::PointLL>& shape);
 
 /**
- * Build an entire valhalla tileset give a config file and some input pbfs
- * @param config              used to tell the function where and how to build the tiles
- * @param input_files         tells what osm pbf files to build the tiles from
- * @param bin_file_prefix     name prefix for mmapped flat files used when parsing the osm pbf
- * @param free_protobuf       whether or not to unload the protobuffer lib, you cant use libpbf
- * after doing this
- *
+ * Build an entire valhalla tileset give a config file and some input pbfs. The
+ * tile building process is split into stages. This method allows either the entire
+ * tile building pipeline to run (default) or a subset of the stages to run.
+ * @param config        Used to tell the function where and how to build the tiles
+ * @param input_files   Tells what osm pbf files to build the tiles from
+ * @param start_stage   Starting stage of the pipeline to run
+ * @param end_stage     End stage of the pipeline to run
  */
 void build_tile_set(const boost::property_tree::ptree& config,
                     const std::vector<std::string>& input_files,
-                    const std::string& bin_file_prefix = "",
-                    bool free_protobuf = true);
+                    const BuildStage start_stage = BuildStage::kParse,
+                    const BuildStage end_stage = BuildStage::kValidate);
 
 } // namespace mjolnir
 } // namespace valhalla